requirements.txt
blinker==1.9.0
click==8.3.0
colorama==0.4.6
Flask==3.0.0
Flask-Cors==4.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==6.0.2
MarkupSafe==3.0.3
PyPDF2==3.0.1
python-docx==1.2.0
python-dotenv==1.0.0
typing_extensions==4.15.0
Werkzeug==3.1.3
